[MP][refactor] move all dataclass to a separate file
