+++ lmcache/integration/vllm/lmcache_mp_connector_0180.py
@@ -8,7 +8,8 @@ from typing import TYPE_CHECKING, Any, Literal
 
 import torch
 import zmq
-from lmcache import torch_dev, torch_device_type
+from lmcache import torch_dev
+from lmcache.integration.vllm.lmcache_mp_metadata import LoadStoreOp
 from lmcache.integration.vllm.utils import mla_enabled
 from lmcache.utils import check_interprocess_event_support, init_logger as lmcache_init_logger
 
@@ -28,13 +29,11 @@ try:
     from lmcache.integration.vllm.vllm_multi_process_adapter import (
         LMCacheMPSchedulerAdapter,
         LMCacheMPWorkerAdapter,
-        LoadStoreOp,
     )
 except ImportError:
     from vllm.distributed.kv_transfer.kv_connector.v1.lmcache_integration import (
         LMCacheMPSchedulerAdapter,
         LMCacheMPWorkerAdapter,
-        LoadStoreOp,
     )
 
 if TYPE_CHECKING:
